Add files via upload
To ensure accessibility for users worldwide and to overcome the language barrier, which can be a major challenge for some, we have translated Walrus documents into Arabic.

We kindly ask you to review the translation first. Additionally, we will conduct a thorough review to ensure accuracy and clarity.

Other translators are also welcome to refine and enhance the translation to maintain the highest quality.

Arabic is one of the most widely spoken languages globally, with a rapidly growing interest in blockchain technology, particularly in countries like the UAE and Saudi Arabia. Given this expansion, we believe this translation is highly valuable and will significantly contribute to the adoption of Walrus in the Arabic-speaking community.